requirements.txt
# Required
django
djangorestframework

# For MySQL
mysql-connector
pymysql
